+++ pwsafe/pwsafe/src/ui/Windows/OptionsBackup.cpp
@@ -207,7 +207,7 @@ BOOL COptionsBackup::OnInitDialog()
 void COptionsBackup::OnComboChanged()
 {
   int nIndex = m_backupsuffix_cbox.GetCurSel();
-  m_backupsuffix = m_backupsuffix_cbox.GetItemData(nIndex);
+  m_backupsuffix = (int)m_backupsuffix_cbox.GetItemData(nIndex);
   if (m_backupsuffix == PWSprefs::BKSFX_IncNumber) {
     GetDlgItem(IDC_BACKUPMAXINC)->EnableWindow(TRUE);
     GetDlgItem(IDC_BKPMAXINCSPIN)->EnableWindow(TRUE);
@@ -428,7 +428,7 @@ void COptionsBackup::OnBrowseForLocation()
     // Create a buffer to store the path, then
     // get the path.
     wchar_t buffer[_MAX_PATH] = { 0 };
-    if(::SHGetPathFromIDList(pIDL, buffer) != 0)
+    if (::SHGetPathFromIDList(pIDL, buffer) != 0)
       m_userbackupotherlocation = CString(buffer);
     else
       m_userbackupotherlocation = L"";
