+++ pwsafe/pwsafe/src/ui/wxWidgets/addeditpropsheet.cpp
@@ -713,7 +713,7 @@ void AddEditPropSheet::ItemFieldsToPropSheet()
   if (!group.empty())
     for (size_t igrp = 0; igrp < aryGroups.size(); igrp++)
       if (group == aryGroups[igrp].c_str()) {
-        m_groupCtrl->SetSelection(igrp);
+        m_groupCtrl->SetSelection((int)igrp);
         break;
       }
